requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
blinker==1.8.2
click==8.1.7
discord.py==2.3.2
Flask==3.0.3
frozenlist==1.4.1
gcloud==0.17.0
googleapis-common-protos==1.63.1
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jws==0.1.3
MarkupSafe==2.1.5
multidict==6.0.5
oauth2client==3.0.0
protobuf==5.27.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.4.3
pyparsing==3.1.2
Pyrebase==3.0.27
python-dotenv==1.0.1
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
typing==3.7.4.3
Werkzeug==3.0.3
yarl==1.9.4
